Modularize request transformer for lint compliance
